requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.1
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
greenlet==3.1.1
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.10
langchain-core==0.3.22
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.147
llama-cloud==0.1.6
llama-index==0.12.2
llama-index-agent-introspective==0.3.0
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.2
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-groq==0.3.0
llama-index-llms-openai==0.3.2
llama-index-llms-openai-like==0.3.0
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.15
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openai==1.56.0
openpyxl==3.1.5
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plotai==0.0.6
plotly==5.24.1
propcache==0.2.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDocX==0.9.10
pydyf==0.11.0
pymongo==4.10.1
pyparsing==3.2.0
pypdf==5.1.0
pyphen==0.15.0
pyppt==0.2.2
pypptx==0.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.5.7
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
XlsxWriter==3.2.0
yarl==1.18.3
zipp==3.21.0
zopfli==0.2.3.post1
